Return metadata information for failed requests.
Requests that got a response used to include some metadata information
in the response headers, such as version, stream id, and latency. This
change also includes that information when there is no response and an
NSError is being returned, by using the same header names as keys in
the userInfo member of NSError.

This change also removes the parallelism metadata, as it wasn't being
used and is available in the configuration.