Track bytes sent and received per request.
Each stream tracks on-the-wire bytes transmitted and received. This
accounts for all SPDY framing and stream-related control messages,
but does not factor in non-stream-related control messages.

These fields are not yet properly exposed to the application, as using
the metadata mechanism (adds to response headers) happens too soon,
before the body has been read and accounted for. tx_bytes should be
accurate, though.