Add "origin endpoint" and proxy configuration.
Connecting to an origin no longer means we simply connect to the origin's
host/port pair. If there is a proxy in between, then the host/port is
different. A SPDYOriginEndpoint represents the actual network host/port
being connected to in order to reach the origin. Also, there may be more
than one proxy that needs to be tried, if the first one fails.

All the code for interfacing with Apple's proxy configuration APIs
(CFNetwork*) is included here.

This code is not used anywhere yet in this change.