Add proxy support in SPDYSocket.
This change adds support for connecting to one or more HTTP proxies. If
a connection fails, the next proxy is tried, and finally the direct
connection is attempted. This all happens internally in SPDYSocket,
and the interface it exposes remains nearly identical in behavior.

Still need to implement proper proxy connection timeouts and clean up
some TODOs.