Unit tests for stream closing iterator bug.
This provides a crude mechanism for mocking the SPDYSocket and
SPDYSession without introducing a lovely dependency like OCMock. That
should be considered for the future, but for now, this works.

The beginnings of a SPDYSessionTest class are here, but it's still
sparse and needs more filling in. A mainline case for the tx_bytes /
rx_bytes / metadata change are here, as is a test case for the
"closing and removing a stream while iterating throws exception" bug.