Resolve auto-config proxy URL using Apple api.
Schedule run loop source on current run loop, and let the caller
provide a completion block to run once resolution is done. Keeping
the correct order of proxies, in case of multiple possibilities, is
simplified by restricting the results to 1 entry. This could change in
the future, but is being kept simple for now.

The whole SPDYOrginEndpointManager implementation for getting and
resolving proxy configs had to move to an asynchronous model,
modifying a list of proxies in an instance variable.

When no proxy auto-config URL is found, then the completion callback
happens immediately on the current thread.