Unit tests for proxy config and socket connect.
Tests for SPDYEndpointManager by mocking the system calls to control
the proxy config that gets returned. Ensure different types work and
multiple proxies work (by only returning a single item).

Tests for connect in SPDYSocket while using a proxy, both HTTPS and
auto-config. Doesn't yet cover all the cases for connection success /
failure, especially around the proxy behavior, but this combined with
the unit tests for SPDYSocketOps gets close. It's really reaching
the point we need to bring in OCMock and/or refactor SPDYSocket.

These tests revealed a few issues in the code that are fixed here.