Add basic metadata to response headers temporarily.
To ease the transition to using the new delegate for metadata, we'll
continue reporting version, stream id, and latency in the response
headers. This maintains parity with the old behavior. This change
should be removed once all the issues have been worked through for
the delegate and its interface locks down.