Fix NSAsserts in _write / _read.
These will fire under normal circumstances, whenever a write or read
operation is in progress and the TCP connection is going away. This
could occur when the server is gracefully shutting things down -- idle
timeout, draining, etc. It's not common, but does happen.

Instead of asserting, we'll log an INFO level since it is still an
interesting action that may prove useful when analyzing logs and
determining behavior.