Support for app retrieving metadata.
Using either the NSURLResponse or the NSError provided from the URL
loading system, the app can retrieve the SPDY metadata at any time until
the final delegate callback is made. This is achieved using an opaque
weak reference and a simple lookup dictionary, and relies on the
strong reference from SPDYProtocol to SPDYStream to SPDYMetadata.