Close remote side when issuing RST_STREAM.
After sending a RST_STREAM, the remote side will never send us another
frame. We need to mark it as closed to allow the stream to close itself
once the local side is marked as closed.

streamCancel worked fine prior to this change, but I refactored it here
a bit to follow a consistent code path.