Fix early release of SPDYMetadata.
Depending on how the NSURL loading system deallocs the protocol,
the SPDYMetadata could be released before the didFinishLoading callback
even executed, or in the middle of it. It seems that iOS 8 releases
the protocol earlier than iOS 7 did.

This change removes all the dictionary code to map identifiers to
SPDYMetadata instances and instead just associates the metadata instance
with the identifier instance. A unique identifier instance is returned
for each associated dictionary it is insert into.

As long as the NSString* identifier is alive, the SPDYMetadata for it
will also be alive.