Fix redirects inside a WebView.
[NSURL initWithString:relativeToURL] produces odd results. The
URLs do not compare unless you examine the absolute strings, and when
going through a WebView on iOS8 the entire host gets stripped off. This
is not the case with iOS7. This fixes it by constructing a second NSURL
for the redirect based on the absolute string.

In addition, a 302 should behave like a 303 in that it rewrites a POST
as a GET. This is undefined in the HTTP spec but many user agents do
this, including it seems Apple's own in its HTTP stack as used by a
WebView. This was observed with Charles proxy, so this change makes
CocoaSPDY behave the same way.

Finally, when redirecting from a POST to a GET, the content-length and
content-type headers need to be stripped, as do all forms of body
content. Those headers are added automatically during canonicalization.