Support for multiple proxies.
Adds support for .pac files to configure multiple proxies, to be
attempted sequentially. Also allows for the addition of a fallback-to-
direct entry, to maintain backwards-compatible behavior when the proxy
connection fails.

We only consider a proxy connection "failed" and move on to the next
attempt in certain circumstances.
- on initial connection during setup of streams, etc
- on tcp connect failure
- any failure sending a CONNECT or receiving the reply
- any reply from the proxy that is not a 2xx

Any other failure is considered an endpoint failure and the socket is
closed with no additional proxy attempts. This is essentially any time
after receiving a 2xx response from the proxy. Either the start of the
TLS session or the sending of the first frame are considered "endpoint
communication".

A special case here is the receipt of a 407 "Proxy authentication
required" response from the proxy. As we don't yet support
authentication, this will result in attempting the next proxy connection.
Since SPDYOriginEndpointManager ensures there is always a direct
connection fallback at the end of the list, we'll always get back-compat
behavior.