Return proxy status in metadata.
Proxy status is used to indicate the state of the proxy configuration,
and in conjunction with the "via proxy" metadata, can be used to
determine the current state of the connection.

For instance:
- viaProxy=false and proxyStatus=Manual means a connection
to the manually-configured proxy failed.

- viaProxy=false and proxyStatus=ManualWithAuth means a connection
to the manuall-configured proxy failed with a 407 response.

- viaProxy=true and proxyStatus=Auto means a .pac file was downloaded
and configured a proxy that we connected to.

- viaProxy=false and proxyStatus=AutoInvalid means a .pac file failed
to download, or returned no valid proxy servers, such as SOCKS or a
typo.