Track detailed stream timings and expose in metadata.
These timings use opaque absolute timestamps. While this makes it a bit
harder to consume, it gives more flexibility and provides more
information. The time between the app starting the request and the
SPDY protocol first receiving it can be measured, for instance. It will
also allow for future additions without breaking the meaning of any
existing timings.