Add richer NSURLSession support.
This change provides SPDY with with NSURLSession, NSURLSessionTask,
and associated configuration. This provides several benefits:

SPDY will have access to the app's delegate and operation queue, which
can be extended and used for additional callbacks, such as informing
the app a request is about to be dispatched to SPDY (implemented here),
or a new pushed stream was received (not yet implemented).

The NSURLSessionConfiguration is available, which Apple failed to
provided via the usual NSURLProtocol interfaces. This means we can
retrieve per-request cookie and cache configuration.

A new SPDYURLSessionDelegate extended delegate is now exposed. The app
may choose to implement this via the NSURLSession.delegate instance. It
will return the SPDYProtocol instance to the app, which can be used now
to directly retrieve the metadata, even in the cases where a NSURL-based
NSError is returned for the request. Previously it was impossible to
tell if CocoaSPDY serviced the request or not in this case (e.g.,
request timeouts).

Note: In order for this to work, the app must set the SPDYURLSession
property on the NSURLRequest instance prior to starting the request.
SPDY uses some indirection, to avoid serialization warnings when
caching the NSURLRequest, to store this reference.

Big thanks to @NSProgrammer for doing most of this work.