Remap kCFErrorDomainCFNetwork to NSURLErrorDomain.
Most apps are expecting NSURLErrors coming from the NSURL system, so
we should provide those when possible. All NSURLErrorDomain error codes
are based on kCFErrorDomainCFNetwork error codes, so that's easy. There
are some kCFErrorDomainCFNetwork codes, however, not present in the
NSURLErrorDomain. Some can be mapped manually; the rest we'll just
pretend like they are, but include the original error in the userInfo.

Also downgraded some log messages to more appropriate levels,
particularly when you consider a phone can be in airplane mode.