Fix rxBytes and rxBodyBytes in metadata.
Data frames were incorrectly reporting their size in the metadata
fields rxBytes and rxBodyBytes. If a data frame got broken up into
multiple TCP segments, those would each be reported to the SPDYSession
as a "data frame", with the encodedLength set to the overall frame
size (for the first segment) or the remaining segment sizes.

It was, essentially, indeterminately wrong from the POV of the app.