Improve session manager failure handling.
When a "fast fail" is detected, which is defined as a socket failing
to initialize or connect when there are no other successfully connected
sockets, we will now throttle the creation of additional sockets by
waiting a small amount of time before the next attempt. Previously,
each HTTP request would  force a new socket creation and connection
attempt. This was potentially dozens of creations, or more, per second.

The delay will only apply to requests queued after a socket fails to
connect. The root error will already have been returned to the app,
during the previous failure. This delay should not be noticeable -- even
if the next socket succeeds, it will be right after a catastrophic
network failure, and the delay will just be lost in the noise.

There was a major bug in _fillSessionPool. If the pool size was > 1, and
if the first SPDYSession (socket) was successfully created but
subsequent ones failed, it would enter an infinite loop until they were
eventually successfully created. This is an unlikely, if not nearly
impossible, scenario in the real world, as I cannot find a way to make
the SPDYSocket fail to initialize the socket except by artificial means.
All failures, even "airplane mode", come asynchronously during a stream
callback.

Also improve socket error messages in SPDYSocket to have unique
messages, and fix case where a stream error could return nil.