Defensive code for SPDYSocket crash.
Inside CFStreamCreatePairWithSocketToHost, generally inside either the
constructor or destructor for HostBase, we sometimes see a crash.
Sometimes it is trying to access address 0x0, sometimes free() detects
corruption.

Let's have no references to self during that call, and let's ensure
the hostname is sufficiently retained.

Also, because it's somewhat related, removing an unnecessary extra
reference to the current endpoint. It's available from the endpoint
manager.