Add support for NSURLCache.
The NSURL loading system has subtle and poorly documented behavior
regarding caching. CocoaSPDY was not doing the right thing, thus
there was no support for NSURLCache in the protocol. This patch
adds basic support for both NSURLConnection and NSURLSession based
requests. This is not yet a fully-featured client caching
implementation.

If the request specifies a NSURLRequest cachePolicy of:

NSURLRequestUseProtocolCachePolicy
- NSURL system does not provide a cached response to the protocol
constructor. It is up to the protocol to load and validate the
cached response.

NSURLRequestReturnCacheDataElseLoad
- NSURL system will provide the cached response, if available. The
protocol is expected to validate the response, and load if not
available or not valid.

NSURLRequestReturnCacheDataDontLoad
- NSURL system will provide the cached response, if available. The
protocol is expected to validate the response. The protocol will not
be loaded if no cached response is available, or if one is but is
invalid, the protocol should not load the request.

In the cases where the protocol has a cached response and it is valid,
it is supposed to call URLProtocol:cachedResponseIsValid. CocoaSPDY
was not doing this. Determining validity of the cached response is the
job of the protocol, and a basic implementation has been provided here.

CocoaSPDY also has to jump through some hoops whenever NSURLSession is
being used, as Apple has not provided a way to get the
NSURLSessionConfiguration and thus we cannot get the right NSURLCache.
Fortunately we have already provided a workaround for this.

SPDYMetadata has been extended to provide the source of the response.