Fix crash in removing the runloop source.
When using a proxy auto-configuration URL, we have to schedule the
execution on a runloop. This patch attempts two fixes, though both
are guesses as I cannot get a repro. The incidents of crashes in the
CFRunLoopRemoveSource() call in _removeRunLoopSource has gone up
dramatically for iOS 10.

1. Simplify run loop modes. Just use common modes.

2. Use CFRunLoopSourceInvalidate instead of CFRunLoopRemoveSource. It
will remove the source from all modes and invalidate it.